+++ gdbserver/netbsd-low.h
@@ -46,7 +46,7 @@ public:
 
   void post_create_inferior () override;
 
-  int attach (unsigned long pid) override;
+  int attach (int pid) override;
 
   int kill (process_info *proc) override;
